% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinstallinfo.R
\name{installinfo}
\alias{installinfo}
\title{Get install information about a package}
\usage{
installinfo(name)
}
\arguments{
\item{name}{A package name}
}
\value{
A data frame
}
\description{
Quickly get LibPath, version, built and package date of a package
}
