% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds.R
\name{seeds}
\alias{seeds}
\title{Generate a set of seeds}
\usage{
seeds(big, small, reps = 1000)
}
\arguments{
\item{big}{An integer to increment over large groups of simulations, such as a study.}

\item{small}{An integer to increment faster than \code{big}, for example over simulations experiments within a study.}
}
\description{
For simulation studies a set of seeds corresponding to each repeat can be useful for reproducibility and investigating rare cases. This function creates that \code{big} and \code{small} useful for separating studies to have different seeds.
}
\details{
Integers are sampled up to \code{.Machine$integer.max} after setting the seed according to \code{big} and \code{small}: \code{big * 1E5 + small}. So for each distinct pair of \code{big} and \code{small}, the set of seeds will be different \code{small < 1E5}.
}
