% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/superheroes.R
\name{batman}
\alias{batman}
\title{Replicate an extremely cheesy theme song.}
\usage{
batman(n)
}
\arguments{
\item{n}{A number.  Number of NaNs desired in the output.}
}
\value{
A Batman-themed list.
}
\description{
Replicate an extremely cheesy theme song.
}

