% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/superheroes.R
\name{batman}
\alias{batman}
\title{Replicate an extremely cheesy theme song.}
\usage{
batman(n)
}
\arguments{
\item{n}{A number.  Restrictions are the same as for the times param of rep.}
}
\value{
A Batman-themed list.
}
\description{
Replicate an extremely cheesy theme song.
}
\examples{
Run it and see for yourself. ;-)
}

